fct_read_data

Pull Data From ActivityInfo

fct_read_data(form)

Value

list with the content of the activity info DB * df5W * dfadmin1 * dfadmin2 * dfindicator * dfpartner

Details

Note that this function is specific to a specific DB schema - the form to query as well as the columns variabl name are hard coded and would need to be adjusted if the DB change

Note also that the function is based on a token stored as an environment variable

# Credentials set up as environment variables -

# token <- "activityinfotoken.." # print(Sys.setenv(ACTIVITYINFOTOKEN = token)) # Sys.getenv("ACTIVITYINFOTOKEN") # rm(token)

Examples

result <- fct_read_data()
## list of results

head(result[["df5W"]], 10)
#>    Country           Admin1 Admin2
#> 1   Mexico         Coahuila   <NA>
#> 2   Mexico  Baja California   <NA>
#> 3   Mexico       Nuevo León   <NA>
#> 4   Mexico Ciudad de México   <NA>
#> 5   Mexico          Chiapas   <NA>
#> 6   Mexico          Tabasco   <NA>
#> 7   Mexico   Aguascalientes   <NA>
#> 8   Mexico  Baja California   <NA>
#> 9   Mexico Ciudad de México   <NA>
#> 10  Mexico Estado de México   <NA>
#>                                            Appealing_org Implementation
#> 1  United Nations High Commissioner for Refugees (UNHCR)             No
#> 2  United Nations High Commissioner for Refugees (UNHCR)             No
#> 3  United Nations High Commissioner for Refugees (UNHCR)             No
#> 4  United Nations High Commissioner for Refugees (UNHCR)             No
#> 5  United Nations High Commissioner for Refugees (UNHCR)             No
#> 6  United Nations High Commissioner for Refugees (UNHCR)             No
#> 7  United Nations High Commissioner for Refugees (UNHCR)             No
#> 8  United Nations High Commissioner for Refugees (UNHCR)             No
#> 9  United Nations High Commissioner for Refugees (UNHCR)             No
#> 10 United Nations High Commissioner for Refugees (UNHCR)             No
#>    Implementing_partner   Month                          Subsector
#> 1                  <NA> 2023-01               Protection (General)
#> 2                  <NA> 2023-01               Protection (General)
#> 3                  <NA> 2023-01               Protection (General)
#> 4                  <NA> 2023-01 Multipurpose Cash Assistance (MPC)
#> 5                  <NA> 2023-01 Multipurpose Cash Assistance (MPC)
#> 6                  <NA> 2023-01 Multipurpose Cash Assistance (MPC)
#> 7                  <NA> 2023-01                          Education
#> 8                  <NA> 2023-01                          Education
#> 9                  <NA> 2023-01                          Education
#> 10                 <NA> 2023-01                          Education
#>                                                                                                  Indicator
#> 1           # of refugees and migrants who received protection-related assistance and specialized services
#> 2           # of refugees and migrants who received protection-related assistance and specialized services
#> 3           # of refugees and migrants who received protection-related assistance and specialized services
#> 4                                      # of individuals benefitting from multipurpose cash transfers (MPC)
#> 5                                      # of individuals benefitting from multipurpose cash transfers (MPC)
#> 6                                      # of individuals benefitting from multipurpose cash transfers (MPC)
#> 7  # of refugees and migrants enrolled in formal educational institutions or non-formal education programs
#> 8  # of refugees and migrants enrolled in formal educational institutions or non-formal education programs
#> 9  # of refugees and migrants enrolled in formal educational institutions or non-formal education programs
#> 10 # of refugees and migrants enrolled in formal educational institutions or non-formal education programs
#>                                        Activity_Name Activity_Description
#> 1  Apoyo para la naturalización (CBI y otros gastos)       Entrega de CBI
#> 2  Apoyo para la naturalización (CBI y otros gastos)       Entrega de CBI
#> 3  Apoyo para la naturalización (CBI y otros gastos)       Entrega de CBI
#> 4                       Asistencia Humanitaria (MPG)       Entrega de CBI
#> 5                       Asistencia Humanitaria (MPG)       Entrega de CBI
#> 6                       Asistencia Humanitaria (MPG)       Entrega de CBI
#> 7                          CBI apoyos universitarios       Entrega de CBI
#> 8                          CBI apoyos universitarios       Entrega de CBI
#> 9                          CBI apoyos universitarios       Entrega de CBI
#> 10                         CBI apoyos universitarios       Entrega de CBI
#>    RMRPActivity CVA Value               Delivery_mechanism Quantity_output
#> 1           Yes Yes   880 Other electronic cash mechanisms              NA
#> 2           Yes Yes    89 Other electronic cash mechanisms              NA
#> 3           Yes Yes  1301 Other electronic cash mechanisms              NA
#> 4           Yes Yes  3416 Other electronic cash mechanisms              NA
#> 5           Yes Yes 12297 Other electronic cash mechanisms              NA
#> 6           Yes Yes   795 Other electronic cash mechanisms              NA
#> 7           Yes Yes  3456 Other electronic cash mechanisms              NA
#> 8           Yes Yes  2973 Other electronic cash mechanisms              NA
#> 9           Yes Yes  6157 Other electronic cash mechanisms              NA
#> 10          Yes Yes  1190 Other electronic cash mechanisms              NA
#>    Total_monthly New_beneficiaries IN_DESTINATION IN_TRANSIT Host_Communities
#> 1              2                 2              2          0                0
#> 2              2                 2              2          0                0
#> 3              4                 4              4          0                0
#> 4             36                36             36          0                0
#> 5            112               112            112          0                0
#> 6              8                 8              8          0                0
#> 7             16                16             16          0                0
#> 8              7                 7              7          0                0
#> 9             25                25             25          0                0
#> 10             6                 6              6          0                0
#>    PENDULARS Returnees Girls Boys Women Men Other_under Other_above
#> 1          0         0     0    0     1   1           0           0
#> 2          0         0     0    0     1   1           0           0
#> 3          0         0     0    0     2   2           0           0
#> 4          0         0     7    7     9  13           0           0
#> 5          0         0    18   22    38  34           0           0
#> 6          0         0     2    2     2   2           0           0
#> 7          0         0     0    0    11   5           0           0
#> 8          0         0     0    0     4   3           0           0
#> 9          0         0     0    0    17   8           0           0
#> 10         0         0     0    0     3   3           0           0
head(result[["dfadmin1"]], 10)
#> # A tibble: 10 × 4
#>    Country   Admin1                          ISOCode countryadmin1              
#>    <chr>     <chr>                           <chr>   <chr>                      
#>  1 Argentina Buenos Aires                    AR-B    Argentina Buenos Aires     
#>  2 Argentina Catamarca                       AR-K    Argentina Catamarca        
#>  3 Argentina Chaco                           AR-H    Argentina Chaco            
#>  4 Argentina Chubut                          AR-U    Argentina Chubut           
#>  5 Argentina Ciudad Autonoma de Buenos Aires AR-C    Argentina Ciudad Autonoma …
#>  6 Argentina Córdoba                         AR-X    Argentina Córdoba          
#>  7 Argentina Corrientes                      AR-W    Argentina Corrientes       
#>  8 Argentina Entre Ríos                      AR-E    Argentina Entre Ríos       
#>  9 Argentina Formosa                         AR-P    Argentina Formosa          
#> 10 Argentina Jujuy                           AR-Y    Argentina Jujuy            
head(result[["dfadmin2"]], 10)
#> # A tibble: 10 × 4
#>    Country   Admin1       Admin2                         admin1and2             
#>    <chr>     <chr>        <chr>                          <chr>                  
#>  1 Argentina Buenos Aires 25 de Mayo (Buenos Aires)      Buenos Aires 25 de May…
#>  2 Argentina Buenos Aires 9 de Julio (Buenos Aires)      Buenos Aires 9 de Juli…
#>  3 Argentina Buenos Aires Adolfo Alsina (Buenos Aires)   Buenos Aires Adolfo Al…
#>  4 Argentina Buenos Aires Adolfo Gonzáles Chaves         Buenos Aires Adolfo Go…
#>  5 Argentina Buenos Aires Alberti                        Buenos Aires Alberti   
#>  6 Argentina Buenos Aires Almirante Brown (Buenos Aires) Buenos Aires Almirante…
#>  7 Argentina Buenos Aires Arrecifes                      Buenos Aires Arrecifes 
#>  8 Argentina Buenos Aires Avellaneda (Buenos Aires)      Buenos Aires Avellaned…
#>  9 Argentina Buenos Aires Ayacucho (Buenos Aires)        Buenos Aires Ayacucho …
#> 10 Argentina Buenos Aires Azul                           Buenos Aires Azul      
head(result[["dfindicator"]], 10)
#> # A tibble: 10 × 5
#>    CODE      Subsector                       Indicator   IndicatorType sectindic
#>    <chr>     <chr>                           <chr>       <chr>         <chr>    
#>  1 RE-CP1    Protection (Child Protection)   # of refug… Direct Assis… Protecti…
#>  2 RE-CP2    Protection (Child Protection)   # of refug… Direct Assis… Protecti…
#>  3 RE-CP3    Protection (Child Protection)   # of indiv… Capacity Bui… Protecti…
#>  4 RE-CP4    Protection (Child Protection)   # of campa… Campaign      Protecti…
#>  5 RE-CSC01  Common Services (Coordination)  # of multi… Other         Common S…
#>  6 RE-CSCO01 Common Services (Communication) # of impre… Other         Common S…
#>  7 RE-CSCO02 Common Services (Communication) # of views… Other         Common S…
#>  8 RE-CSCO03 Common Services (Communication) # of commu… Other         Common S…
#>  9 RE-CSCO04 Common Services (Communication) # of visit… Other         Common S…
#> 10 RE-CSCO05 Common Services (Communication) # of insti… Capacity Bui… Common S…
head(result[["dfpartner"]], 10)
#>    AOIDORG                                           Name Type RMlead
#> 1   ORG001                     100% Diversidad y Derechos NNGO   <NA>
#> 2   ORG002                                          ACAPS INGO   <NA>
#> 3   ORG003                                          ACTED INGO   <NA>
#> 4   ORG004                          Action against Hunger INGO   <NA>
#> 5   ORG005                                      ActionAid INGO   <NA>
#> 6   ORG006 Adventist Development and Relief Agency (ADRA) INGO   <NA>
#> 7   ORG007                                   AID FOR AIDS INGO   <NA>
#> 8   ORG008               AIDS Healthcare Foundation (AHF) INGO   <NA>
#> 9   ORG009                     Alas de Colibrí Foundation NNGO   <NA>
#> 10  ORG010                          Americares Foundation INGO   <NA>